requirements_base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0

# Django
django==1.10.8

# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Forms
django-braces==1.12.0
django-crispy-forms==1.6.0

# Models
django-model-utils==3.1.1

# Images
Pillow==5.0.0


# Python-PostgreSQL Database Adapter
psycopg2-binary==2.7.4

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2018.3

# Your custom requirements go here

# Support for rendering Markdown
django-markdown-deux==1.0.5

# Django-rules for permission checking
rules==1.3

# File Storage
django-db-file-storage==0.5.0

# Multi-file upload
django-multiupload==0.5.2

# Django-REST-Framework and friends
djangorestframework==3.7.7
Markdown==2.6.11
django_filter==1.1.0
rest-framework-generic-relations==1.1.0
django-rest-knox==3.1.3
dry-rest-permissions==0.1.10

# Django form wizards
django-formtools==2.1

# Page range
pagerange==0.4
